chore: improve Result<T, Err> Types
