+++ packages/type-utils/src/index.ts
@@ -1,7 +1,8 @@
 export * from './array';
+export * from './branded';
+export * from './exhaustive';
+export * from './object';
 export * from './overloads';
+export * from './result';
 export * from './timeout';
 export * from './utils';
-export * from './object';
-export * from './exhaustive';
-export * from './branded';
