+++ packages/type-utils/src/utils.ts
@@ -143,45 +143,3 @@ export type FilterOutFromUnionByTypeProperty<
         ? never
         : { [K in KeyName]: Exclude<ActualValue, ValueToExclude> } & Omit<Union, KeyName>
     : Union;
-
-/**
- * Success type
- *
- * Example:
- *  ```
- *  type T = OK<string>;
- *  const t: T = 'value';
- *  ```
- */
-export type OK<S> = {
-    ok: true;
-    value: S;
-};
-
-/**
- * Error type
- *
- * Example:
- *  ```
- *  type T = Err<number>;
- *  const t: T = 1;
- *  ```
- */
-export type Err<E> = {
-    ok: false;
-    error: E;
-};
-
-/**
- * Result type
- *
- * Example:
- *  ```
- *  type T = Result<string, number>;
- *  const t: T = 'value';
- *  ```
- */
-export type Result<S, E> = OK<S> | Err<E>;
-
-export const ok = <S>(value: S): OK<S> => ({ ok: true, value });
-export const err = <E>(error: E): Err<E> => ({ ok: false, error });
